Fix CJK/rotated glyph spacing and centering in vertical text
- Use measureText() for every run's advance (including rotated CJK) instead
  of the font line height, so CJK characters are no longer over-spaced.
- Center rotated (CJK/emoji) glyphs across the line and align them to the top
  (-fm.top); vertically center non-rotated glyphs within the line.
- Use float for layout positions/metrics for sub-pixel accuracy.
- Simplify MongolTextLine.measure() by dropping the rotated special-case.

+++ mongol-library/src/main/java/net/studymongolian/mongollibrary/MongolLayout.java
@@ -128,10 +128,10 @@ public class MongolLayout {
 
         if (needsLineUpdate) updateLines();
 
-        int metricsBottom = mTextPaint.getFontMetricsInt().bottom;
+        float metricsBottom = mTextPaint.getFontMetricsInt().bottom;
 
-        int x = metricsBottom; // start position of each vertical line
-        int y = 0; // baseline
+        float x = metricsBottom; // start position of each vertical line
+        float y = 0; // baseline
         MongolTextLine tl = MongolTextLine.obtain();
 
         boolean needMultiply = (mSpacingMult != 1 || mSpacingAdd != 0);
